+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 - TabbedPanelLayout.Builder.includeDetailTabPane() renamed includeDetailTabbedPane().
 - EntityPanel, edit window location now relative to table panel, if one is available.
 - EntityEditPanel bug fixed, did not validate before insert and update.
+- EntityTablePanel.createSouthToolBar() renamed createToolBar().
 ### is.codion.swing.framework.server.monitor
 - ServerMonitor and ServerMonitorPanel, tables migrated from plain JTable to FilteredTable.